requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
fpdf2==2.7.5

